Tests/LibWeb/Text/expected/secure-context.txt

{
  "SubtleCrypto": false,
  "Clipboard": false,
  "CookieChangeEvent": true
}
